+++ example/android/app/src/main/kotlin/com/example/thai_id_card_numbers_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.thai_id_card_numbers_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()